Split test_url.py & test_wikidata.py from test_zotero.py
